move resource server to common.separate configuration for LDAP and DB
