+++ jwt-aouth-server/src/main/resources/application.properties
@@ -5,6 +5,7 @@ spring.datasource.username=sa
 spring.datasource.password=
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 logging.level.org.springframework.*: DEBUG
-authentication.mode=DB
+spring.profiles.active=db
 authentication.mode.ldap.activeDirectoryDomain=flytxt.com
 authentication.mode.ldap.activeDirectoryUrl=ldap://192.168.125.10:389
+
