--- resource-server-config/src/main/resources/application.yml
@@ -1,2 +1,0 @@
-server:
-  port: 9001
